try commit buoi2 3

+++ Buoi1/Bai2, 3, 4/Bai2_CauA.cpp
@@ -12,7 +12,7 @@ struct Student
 {
 	string Name;
 	string ID;
-	int DiemTrungBinh;
+	double DiemTrungBinh;
 	Student *SNext;
 };
 
@@ -27,7 +27,7 @@ void initClass(Class &MyClass)
 	MyClass.SHead = MyClass.STail= NULL;
 }
 
-Student *createStudent(string Name, string ID, int Dtb)
+Student *createStudent(string Name, string ID, double Dtb)
 {
 	Student *student = new Student;
 	if (student == NULL)
@@ -65,7 +65,7 @@ void inputClass(Class &MyClass, int &n)
 	for (int i = 0; i < n; i++)
 	{
 		string Name, ID;
-		int Dtb;
+		double Dtb;
 		printf("*Number %d:\n", i);
 		printf("Name: ");
 		cin >> Name;
@@ -77,6 +77,7 @@ void inputClass(Class &MyClass, int &n)
 		addClass(MyClass, sv);
 	}
 }
+
 void outputClass(Class MyClass)
 {
 	printf("! Check class: \n");
