+++ Buoi1/Bai2, 3, 4/Bai2_CauC.cpp
@@ -12,7 +12,7 @@ struct Student
 {
 	string Name;
 	string ID;
-	int DiemTrungBinh;
+	double DiemTrungBinh;
 	Student *SNext;
 };
 
@@ -20,14 +20,14 @@ struct Class
 {
 	Student *SHead;
 	Student *STail;
-};	
+};
 
 void initClass(Class &MyClass)
 {
-	MyClass.SHead = MyClass.STail= NULL;
+	MyClass.SHead = MyClass.STail = NULL;
 }
 
-Student *createStudent(string Name, string ID, int Dtb)
+Student *createStudent(string Name, string ID, double Dtb)
 {
 	Student *student = new Student;
 	if (student == NULL)
@@ -66,7 +66,7 @@ void inputClass(Class &MyClass, int &n)
 	for (int i = 0; i < n; i++)
 	{
 		string Name, ID;
-		int Dtb;
+		double Dtb;
 		printf("*Number %d:\n", i);
 		printf("Name: ");
 		cin >> Name;
@@ -86,7 +86,7 @@ void outputClass(Class MyClass)
 	for (Student *i = MyClass.SHead; i != NULL; i = i->SNext)
 	{
 		printf("*Number %d: ", count);
-		cout <<"Hoc sinh: "<< i->Name <<" MSSV: "<< i->ID <<" Diem: "<< i->DiemTrungBinh<<";"<<endl;
+		cout << "Hoc sinh: " << i->Name << " MSSV: " << i->ID << " Diem: " << i->DiemTrungBinh << ";" << endl;
 		count++;
 	}
 }
